fix l:r slices where l < 0 and relative r >= -l
parse_slice() stored l < 0 as a negative bit offset which was then added
to a relative r offset (that is, r value preceded by unary + operator).
basically, the parsing of "-4:+4" string produced a slice structure with
slice->l = -32 (i.e. 4 bytes) and slice->r = -32 + 32 = 0. consequently,
bits_of_slice() treated the parsed slice as "6:0" and returned no bits
because l > r (6 > 0). this is incorrect behavior as the string "-4:+4"
is supposed to represent a slice of the last 4 input bytes (or 32 bits).